- split CoreGraphics out into it's own framework project. AppKit still embeds CoreGraphics but it is also built standalone on Foundation. (don't link against both) - new experimental AppKit-MacOS-i386 target which sits on Apple's Foundation, ApplicationServices (not operational). Header&api usage fixes to clean this up. - Redid NSBezierPath so it doesn't rely on private Cocotron CoreGraphics API. - Change the NSRaise macros/functions to be inline so AppKit can just include it instead of depending on them in Foundation since they aren't present in OSX. - split NSRaiseException() into it's own Foundation private header and removed use from AppKit. - changed use of NSRangeEntries so AppKit can embed it on MacOS - switched NSOutlineView use of NSInt*CallBacks to NSInteger*CallBacks - added CGPDFContext and CGDataConsumer

+++ AppKit/NSBrowser.h
@@ -48,7 +48,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 -(NSColor *)backgroundColor;
 -(SEL)doubleAction;
 -(Class)matrixClass;
--(int)maxVisibleColumns;
+-(NSInteger)maxVisibleColumns;
 -(BOOL)hasHorizontalScroller;
 -(BOOL)separatesColumns;
 -(BOOL)isTitled;
